Split tidy packages for readability
