+++ assets/js/base/components/cart-checkout/shipping-rates-control-package/index.tsx
@@ -5,60 +5,17 @@ import classNames from 'classnames';
 import { _n, sprintf } from '@wordpress/i18n';
 import { decodeEntities } from '@wordpress/html-entities';
 import type { ReactElement } from 'react';
-import type { PackageRateOption } from '@woocommerce/type-defs/shipping';
 import { Panel } from '@woocommerce/blocks-checkout';
 import Label from '@woocommerce/base-components/label';
 import { useSelectShippingRate } from '@woocommerce/base-context/hooks';
-import type { CartShippingPackageShippingRate } from '@woocommerce/type-defs/cart';
 import { sanitizeHTML } from '@woocommerce/utils';
 
 /**
  * Internal dependencies
  */
 import PackageRates from './package-rates';
+import type { PackageProps } from './types';
 import './style.scss';
-
-interface PackageItem {
-	name: string;
-	key: string;
-	quantity: number;
-}
-
-interface Destination {
-	address_1: string;
-	address_2: string;
-	city: string;
-	state: string;
-	postcode: string;
-	country: string;
-}
-
-export interface PackageData {
-	destination: Destination;
-	name: string;
-	shipping_rates: CartShippingPackageShippingRate[];
-	items: PackageItem[];
-}
-
-export type PackageRateRenderOption = (
-	option: CartShippingPackageShippingRate
-) => PackageRateOption;
-
-// A flag can be ternary if true, false, and undefined are all valid options.
-// In our case, we use this for collapsible and showItems, having a boolean will force that
-// option, having undefined will let the component decide the logic based on other factors.
-export type TernaryFlag = boolean | undefined;
-interface PackageProps {
-	/* PackageId can be a string, WooCommerce Subscriptions uses strings for example, but WooCommerce core uses numbers */
-	packageId: string | number;
-	renderOption?: PackageRateRenderOption | undefined;
-	collapse?: boolean;
-	packageData: PackageData;
-	className?: string;
-	collapsible?: TernaryFlag;
-	noResultsMessage: ReactElement;
-	showItems?: TernaryFlag;
-}
 
 export const ShippingRatesControlPackage = ( {
 	packageId,
@@ -128,20 +85,19 @@ export const ShippingRatesControlPackage = ( {
 			) }
 		</>
 	);
-	const body = (
-		<PackageRates
-			className={ className }
-			noResultsMessage={ noResultsMessage }
-			rates={ packageData.shipping_rates }
-			onSelectRate={ ( newShippingRateId ) =>
-				selectShippingRate( newShippingRateId, packageId )
-			}
-			selectedRate={ packageData.shipping_rates.find(
-				( rate ) => rate.selected
-			) }
-			renderOption={ renderOption }
-		/>
-	);
+
+	const packageRatesProps = {
+		className,
+		noResultsMessage,
+		rates: packageData.shipping_rates,
+		onSelectRate: ( newShippingRateId: string ) =>
+			selectShippingRate( newShippingRateId, packageId ),
+		selectedRate: packageData.shipping_rates.find(
+			( rate ) => rate.selected
+		),
+		renderOption,
+	};
+
 	if ( shouldBeCollapsible ) {
 		return (
 			<Panel
@@ -153,10 +109,11 @@ export const ShippingRatesControlPackage = ( {
 				initialOpen={ false }
 				title={ header }
 			>
-				{ body }
+				<PackageRates { ...packageRatesProps } />
 			</Panel>
 		);
 	}
+
 	return (
 		<div
 			className={ classNames(
@@ -165,7 +122,7 @@ export const ShippingRatesControlPackage = ( {
 			) }
 		>
 			{ header }
-			{ body }
+			<PackageRates { ...packageRatesProps } />
 		</div>
 	);
 };
